requirements.txt
# Framework Web
fastapi==0.116.2
uvicorn==0.35.0

# Validação de Dados
pydantic==2.11.9
pydantic_core==2.33.2
pydantic-settings==2.1.0

# Inteligência Artificial
google-generativeai==0.8.5

# Banco de Dados
pymongo==4.6.1
motor==3.3.2

# HTTP Client
httpx==0.27.0

# Configurações
python-dotenv==1.1.1

# Dependências do FastAPI
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
colorama==0.4.6
h11==0.16.0
idna==3.10
sniffio==1.3.1
starlette==0.48.0
typing-inspection==0.4.1
typing_extensions==4.15.0

# Dependências do Google AI
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.2
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.182.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.75.0
grpcio-status==1.71.2
httplib2==0.31.0
proto-plus==1.26.1
protobuf==5.29.5
pyasn1==0.6.1
pyasn1_modules==0.4.2
pyparsing==3.2.4
requests==2.32.4
rsa==4.9.1
uritemplate==4.2.0
urllib3==2.5.0
wrapt==1.15.0